+++ src/main/kotlin/no/nav/familie/ef/søknad/api/dto/søknadsdialog/Person.kt
@@ -2,5 +2,5 @@ package no.nav.familie.ef.søknad.api.dto.søknadsdialog
 
 import javax.validation.Valid
 
-data class Person(val barn: List<Barn>? = null,
+data class Person(val barn: List<Barn>,
                   @field:Valid val søker: Søker)
